refactor(comms): split quota-pause modules to satisfy sonar size limits
The deploy workflow's `validate` runs `lint:sonar`, which the PR-gate
`realmode` check does not — so #368 merged with 5 SonarJS violations
(files >50 lines, functions >25) and the develop deploy failed at build.

Extract the additions into focused modules, no behaviour change:
- resend/quota.ts        — quotaKindFromName + readQuotaKind (out of response.ts)
- resend/batch-classify.ts — parseIds + classifyBatch (out of batch.ts)
- dispatch/pause-gate.ts  — isPaused + mayDispatch (out of scheduled.ts)
- dispatch/run-summary.ts — summarize (out of run.ts)

lint:sonar 0 errors, full `validate` green, 289 comms-worker tests pass.
