+++ services/comms-worker/src/dispatch/scheduled.ts
@@ -1,7 +1,8 @@
 import { matchesTick } from '../cron/matcher'
 import { logEvent } from '../log/structured'
-import { createSettingsRepo, type SettingsRepo } from '../settings/repo'
+import { createSettingsRepo } from '../settings/repo'
 import { buildRuntimeDeps } from './build-deps'
+import { mayDispatch } from './pause-gate'
 import { runDispatch } from './run'
 import type { DispatchEnv } from './runtime-env'
 import type { DispatchSummary } from './types'
@@ -21,27 +22,6 @@ export type HandleScheduledOptions = {
 
 const defaultDispatcher: Dispatcher = (env, tickAt) =>
   runDispatch(buildRuntimeDeps(env, tickAt))
-
-/**
- * Whether this tick is still inside a quota pause. A pause set on a
- * `daily_quota_exceeded` holds every tick until the quota resets; once
- * `tickAt` reaches it the pause is dropped and the tick proceeds, so
- * the deferred recipients replay on the first tick after the reset.
- * @param repo Settings repo bound to this tick's DB.
- * @param tickAt The tick moment.
- * @returns True when the dispatch is still paused (skip this tick).
- */
-const isPaused = async (
-  repo: SettingsRepo,
-  tickAt: Date
-): Promise<boolean> => {
-  const until = await repo.getPausedUntil()
-  if (until === undefined) return false
-  const untilMs = Date.parse(until)
-  if (Number.isFinite(untilMs) && tickAt.getTime() < untilMs) return true
-  await repo.clearPausedUntil()
-  return false
-}
 
 /**
  * Cron handler: load the saved schedule, decide whether this tick is
@@ -70,10 +50,7 @@ export const handleScheduled = async (
     timezone: sched.timezone,
   })
   if (!matched) return undefined
-  if (await isPaused(settings, tickAt)) {
-    logEvent('tick.paused', { until: await settings.getPausedUntil() })
-    return undefined
-  }
+  if (!(await mayDispatch(settings, tickAt))) return undefined
   const run = opts.dispatcher ?? defaultDispatcher
   return run(env, tickAt)
 }
