+++ services/comms-worker/src/resend/batch.ts
@@ -1,11 +1,6 @@
+import { classifyBatch } from './batch-classify'
 import { RESEND_BATCH_URL } from './batch-request'
-import {
-  DEFAULT_BACKOFF_MS,
-  isRetryableStatus,
-  type QuotaKind,
-  readQuotaKind,
-  retryAfterMs,
-} from './response'
+import { DEFAULT_BACKOFF_MS, type QuotaKind } from './response'
 
 /** Discriminated outcome of one Resend batch round-trip. */
 export type BatchVerdict =
@@ -18,31 +13,6 @@ export type BatchVerdict =
       readonly quota?: QuotaKind
     }
   | { readonly kind: 'fail'; readonly error: string }
-
-const parseIds = (body: unknown): ReadonlyArray<string> => {
-  const data = (body as { data?: unknown } | undefined)?.data
-  if (!Array.isArray(data)) return []
-  return data.map(e => {
-    const id = (e as { id?: unknown } | null)?.id
-    return typeof id === 'string' ? id : ''
-  })
-}
-
-const classifyBatch = async (res: Response): Promise<BatchVerdict> => {
-  if (res.ok) {
-    const body = await res.json().catch(() => undefined)
-    return { kind: 'ok', ids: parseIds(body) }
-  }
-  if (!isRetryableStatus(res.status))
-    return { kind: 'fail', error: `resend ${res.status}` }
-  const quota = res.status === 429 ? await readQuotaKind(res) : undefined
-  return {
-    kind: 'retry',
-    waitMs: retryAfterMs(res),
-    status: res.status,
-    ...(quota ? { quota } : {}),
-  }
-}
 
 const attemptBatch = async (
   doFetch: typeof fetch,
